#816 Removed Manhattan. We will offer only hierarchical layouts from now on. Manhattan was the only non-hierarchical layout and never really used (or of any use; it was just the historically first layout). The distinction of flat and hierarchical layouts introduced only unnecessary complexity.
