--- Assets/SEE/Game/City/AbstractSEECityExtension.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 621ba49b55c556a488ac17930f2177e0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
