--- Assets/SEE/Layout/NodeLayouts/FlatNodeLayout.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8b085be28ad87fc4081c2d477ea44f39
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
