--- Assets/SEE/Layout/NodeLayouts/HierarchicalNodeLayout.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e69bc58071bdb2042bd9eae237a5a19f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
