--- Assets/SEE/Layout/NodeLayouts/ManhattanLayout.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cb4dcca55f39e744da791f93c2af2d68
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
